+++ Sources/PocketFramework/Worksheet/WorksheetRowView.swift
@@ -63,7 +63,7 @@ extension PocketFramework {
                     self.isAdjustmentsEdit.toggle()
                 })
         }
-        public func formatToCurrency(decValue:Decimal) -> String {
+        public func formatToCurrency(decValue:Double) -> String {
             let formatter = NumberFormatter()
             formatter.numberStyle = .currency
             formatter.usesGroupingSeparator = true
